sort scores by score desc with pinia getters
- pinia getters
Getters are exactly the equivalent of computed values for the state of a Store.
They can be defined with the getters property in defineStore().
They receive the state as the first parameter to encourage the usage of `arrow function`:
https://pinia.vuejs.org/core-concepts/getters.html

- access other getters
As with computed properties, you can combine multiple getters. Access any other getter via `this`.
https://pinia.vuejs.org/core-concepts/getters.html#accessing-other-getters